Added configuration definitions.
For the time being, this is very simple and just based off of the default configuration. This also ensures the default configuration is unlinked, to make sure that there are no major conflicts.
In the future, this should receive a dict of options, so that you can define a playbook of different sites. For the time being, I am rushing, so all that this does is create a single site with a very very basic configuration that will conflict if another site config were to be added.