playbook definition (wip)
here we actually define how the two roles will be working together. since we have a location defined in the `clone.yml` in the `glowing-bear` role, we just plug that in for the `nginx` role root directory and name variables. eventually this will be dict'ified, as mentioned earlier, which should make the `nginx` role pretty expandable! (with more work.)